refactor: move lake callbacks to dedicate class
